requirements.txt
# Core dependencies
pyyaml>=6.0
pytest>=8.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0

# LLM API clients
openai>=1.12.0
anthropic>=0.20.0

# Optional: for async HTTP
aiohttp>=3.9.0
httpx>=0.26.0
